requirements.txt
arrow==0.12.1
boto3==1.9.86
botocore==1.12.86
cffi==1.12.3
dateparser==0.7.0
editdistance==0.3.1
fire==0.2.1
flake8==3.7.7
fuzzywuzzy==0.17.0
h5py==2.9.0
img2pdf==0.3.3
ipython==5.8.0
joblib==0.14.1
Keras==2.2.4
keras_tqdm==2.0.1
matplotlib==2.2.3
multiprocessing_logging
numba==0.49.1
numpy==1.16.4
nvgpu>=0.4
opencv-python==3.4.5.20
pandas>=0.23.1
Pillow==5.3.0
# for serialization in tensorpack
pyarrow==0.12.0
pyinterval==1.2.0
python-dateutil==2.7.5
python-Levenshtein==0.12.0
# needs libmagic1 DEB
python-magic==0.4.15
# for tensorpack, needs libcap-dev via apt
python-prctl==1.7; 'linux' in sys_platform
requests==2.21.0
rollbar==0.14.5
scipy==1.1.0
six
scikit-image==0.14.2
scikit-learn==0.20.0
# interactive only
tarjan==0.2.3.2
tensorpack==0.9.4
tqdm==4.23.4
Unidecode==0.4.21
XlsxWriter==1.0.5
imgaug==0.2.8

# --- unit testing ---

jsondiff==1.1.2
jsonschema
mock>=2.0.0
moto<1.4

pytest
pytest-cov
pytest-xdist==1.29.0

ipympl
pympler
click
triarray
Pyqtree
ipywidgets
jupyterlab==1.0.10
tensorflow-gpu==1.13.2
